Add base class BaseAutonomous and Vuforia related stuff
